pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarm-cli"
version = "0.1.0"
description = "SWARM CLI - Track your startup progress with your cohort"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
]

[project.scripts]
swarm = "swarm.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["swarm"]
